Add bottom margin if content is fitting system ui
If the content is fitting the system windows the snackbar gets animated behind the navigation bar. Therefore check if the measured height equals the bottom position of the content view. If it does add a the navigation bar height as bottom margin.

This is rough workaround, further checks need to be added if required.

This might resolve issue #41.