Make RecyclerView a provided dependency
If a developer doesn't use it, then this becomes an unnecessary extra dependency that exists for the sole purpose of attaching to and dismissing when a recyclerview scrolls. By marking it provided (and moving some code around), we can make sure this there for the developers that need it without unnecessarily burdening other developers with it as a transitive dependency.