Fix dismissByReplace jarring animation
Separated the mAnimated field of Snackbar into 2 fields: mShowAnimated
and mDismissAnimated with each field having its own getter and setter
methods. withAnimation(boolean) will change both of these values and
isAnimated() will return true only if both of them are set to true.

In SnackBarManager when replacing a snackbar, dimissAnimation(false) is
called on the previous snackbar and showAnimation(false) is called on
the new snackbar. This provided much smoother visuals when dismissing by
replace.